requirements.txt
tensorflow==2.8.0
protobuf==3.20.3
keras==2.8.0
numpy>=1.19.2
opencv-python>=4.5.1.48
matplotlib>=3.3.4
pillow>=8.1.0
streamlit==1.10.0
scikit-image>=0.18.1
altair==4.2.2
